Fix a bug in the peer discovery
After the new nodes were added the call to #use used #keyspace, which samples a connection and sees what keyspace its in -- but since we just added new nodes that could mean we asked one of the new nodes, and that way all old connections changed keyspace instead.

Don't add the new connections until they're in the right keyspace.