Strategize the peer selection
Extract part of the peer selection from PeerDiscovery. PeerDiscovery still filters out already connected peers, but then a connection strategy takes over the decision process. The default is to only connect to peers in the same data centers as the seeds, and this strategy is now created in Client as part of the connection setup.