Change connection manager to use immutable list
This way, locks are only required around access to the instance variable
and not when modifying the list of connections. Given that the list of
connections is expected to stay relatively stable, this should optimize
the common case.

This was already encapsulated by the snapshot method, so the remaining
methods were rewritten to use that.